mind/requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
cattrs==1.1.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
coverage==5.3.1
cycler==0.10.0
deepspeech==0.9.3
Flask==1.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.3.3
-e git+git@github.com:jonaslimads/robot.git@102379f8455f4541fc2dfea3c7bbb3ea149be115#egg=mind&subdirectory=mind
mypy==0.790
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.8
paho-mqtt==1.5.1
pathspec==0.8.1
Pillow==8.1.0
Pint==0.16.1
pluggy==0.13.1
py==1.10.0
PyAudio==0.2.11
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.11.0
python-dateutil==2.7.5
pytz==2020.5
PyYAML==3.13
regex==2020.11.13
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.2.19
toml==0.10.2
tornado==6.1
tqdm==4.56.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
webrtcvad==2.0.10
websocket-client==0.57.0
Werkzeug==1.0.1
